Adding support for binding - zat server
Adds support to allow binding to a specific address when using zat server. Specifically in my case, this is best used when we use vagrant/virtual box to develop our apps with as we tend to bind them to 0.0.0.0 and setup tools in vagrant to allow us to connect to a port on the virtual machine.